database,seeder modified

+++ database/migrations/2021_09_06_171488_create_districts_table.php
@@ -13,18 +13,25 @@ class CreateDistrictsTable extends Migration
      */
     public function up()
     {
-        Schema::create('districts', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('bn_name')->nullable();
-            $table->string('lat')->nullable();
-            $table->string('lon')->nullable();
-            $table->string('url')->nullable();
-            $table->boolean('status')->default(1);
-            $table->foreignId('division_id')->constrained('divisions')->onDelete('cascade');
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('districts')) {
+            Schema::create('districts', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('bn_name')->nullable();
+                $table->string('lat')->nullable();
+                $table->string('lon')->nullable();
+                $table->string('url')->nullable();
+                $table->boolean('status')->default(1)->index()->comment('1 = active, 0 = inactive');
+                $table->foreignId('division_id')->constrained('divisions')->onDelete('cascade');
+                $table->softDeletes();
+                $table->timestamps();
+
+
+                // Additional indexes
+                $table->index('division_id');
+                $table->index('name');
+            });
+        }
     }
 
     /**
